# docs/swagger.yaml
basePath: /
definitions:
  docs.NullBool:
    properties:
      bool:
        type: boolean
      valid:
        type: boolean
    type: object
  docs.NullFloat64:
    properties:
      float64:
        type: number
      valid:
        type: boolean
    type: object
  docs.NullInt64:
    properties:
      int64:
        type: integer
      valid:
        type: boolean
    type: object
  docs.NullString:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  docs.NullTime:
    properties:
      time:
        type: string
      valid:
        type: boolean
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.AMPPResponse:
    properties:
      co_apid:
        type: integer
      co_disccd:
        $ref: '#/definitions/docs.NullInt64'
      co_indfarmpopcd:
        $ref: '#/definitions/docs.NullInt64'
      co_legal_catcd:
        $ref: '#/definitions/docs.NullInt64'
      co_primarypck:
        $ref: '#/definitions/docs.NullInt64'
      co_secondarypck:
        $ref: '#/definitions/docs.NullInt64'
      co_seq_id:
        type: integer
      co_vppid:
        type: integer
      ds_subp:
        $ref: '#/definitions/docs.NullString'
      dt_discdt:
        $ref: '#/definitions/docs.NullInt64'
      dt_regpublic:
        $ref: '#/definitions/docs.NullInt64'
      no_nm:
        type: string
      nu_appid:
        type: string
      nu_ean13a:
        $ref: '#/definitions/docs.NullString'
      nu_ean13b:
        $ref: '#/definitions/docs.NullString'
      nu_ean13c:
        $ref: '#/definitions/docs.NullString'
      nu_sanreg:
        $ref: '#/definitions/docs.NullInt64'
      st_change_approve:
        type: string
      st_combpackcd:
        $ref: '#/definitions/docs.NullBool'
      st_hosp:
        $ref: '#/definitions/docs.NullBool'
      st_registro_ativo:
        type: string
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.DCBResponse:
    properties:
      co_cas:
        $ref: '#/definitions/docs.NullString'
      co_seq_id:
        type: integer
      ds_classif:
        $ref: '#/definitions/docs.NullString'
      ds_dcb:
        type: string
      nu_dcb:
        type: integer
      st_change_approve:
        type: string
      st_registro_ativo:
        type: string
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.IngredientResponse:
    properties:
      co_dcbcd:
        $ref: '#/definitions/docs.NullInt64'
      co_isrcid:
        $ref: '#/definitions/docs.NullInt64'
      co_ncas:
        $ref: '#/definitions/docs.NullString'
      co_seq_id:
        type: integer
      dt_isiddt:
        $ref: '#/definitions/docs.NullInt64'
      no_nm:
        type: string
      no_nm_pt_br:
        $ref: '#/definitions/docs.NullString'
      nu_isid:
        type: string
      st_change_approve:
        type: string
      st_registro_ativo:
        type: string
    type: object
  github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse:
    properties:
      cursor:
        type: string
      items: {}
      limit:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.VMPPResponse:
    properties:
      co_qty_uomcd:
        $ref: '#/definitions/docs.NullInt64'
      co_qty_uomcd_fornecimento:
        $ref: '#/definitions/docs.NullInt64'
      co_seq_id:
        type: integer
      co_vpid:
        type: integer
      no_nm:
        type: string
      nu_vppid:
        type: string
      qt_qtyval:
        $ref: '#/definitions/docs.NullFloat64'
      qt_qtyval_fornecimento:
        $ref: '#/definitions/docs.NullFloat64'
      st_change_approve:
        type: string
      st_combpackcd:
        $ref: '#/definitions/docs.NullBool'
      st_registro_ativo:
        type: string
    type: object
...
  github_com_silasrm_api-obm_internal_interface_http_dto.VTMResponse:
    properties:
      au_audit_change_datetime:
        $ref: '#/definitions/docs.NullTime'
      au_audit_change_username:
        type: string
      au_audit_create_datetime:
        $ref: '#/definitions/docs.NullTime'
      au_audit_create_username:
        type: string
      co_seq_id:
        type: integer
      no_abbrevnm:
        type: string
      no_nm:
        type: string
      nu_vtmid:
        type: integer
      st_change_approve:
        type: string
      st_registro_ativo:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API do Observatorio de Medicamentos - Brasil (OBM). Gerencia dados
    de medicamentos seguindo o padrao dm%d adaptado para o Brasil.
  title: OBM API
  version: "1.0"
paths:
...
  /api/v1/ampp:
    get:
      description: Lista Actual Medicinal Product Packs com paginacao e filtros
      parameters:
      - default: 20
        description: Limite por pagina
        in: query
        name: limit
        type: integer
      - description: Cursor de paginacao
        in: query
        name: cursor
        type: string
      - description: Filtro por nome (ILIKE)
        in: query
        name: nome
        type: string
      - description: Filtro por status ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar AMPPs
      tags:
      - AMPP
  /api/v1/ampp/{id}:
    get:
      description: Retorna um Actual Medicinal Product Pack pelo seu ID
      parameters:
      - description: ID do AMPP
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.AMPPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter AMPP por ID
      tags:
      - AMPP
  /api/v1/dcb:
    get:
      description: Lista Denominacoes Comuns Brasileiras com paginacao e filtros
      parameters:
      - default: 20
        description: Limite por pagina
        in: query
        name: limit
        type: integer
      - description: Cursor de paginacao
        in: query
        name: cursor
        type: string
      - description: Filtro por nome (ILIKE)
        in: query
        name: nome
        type: string
      - description: Filtro por status ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar DCBs
      tags:
      - DCB
  /api/v1/dcb/{id}:
    get:
      description: Retorna uma Denominacao Comum Brasileira pelo seu ID
      parameters:
      - description: ID do DCB
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.DCBResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter DCB por ID
      tags:
      - DCB
...
  /api/v1/ingredients:
    get:
      description: Lista Ingredient Substances com paginacao e filtros
      parameters:
      - default: 20
        description: Limite por pagina
        in: query
        name: limit
        type: integer
      - description: Cursor de paginacao
        in: query
        name: cursor
        type: string
      - description: Filtro por nome (ILIKE)
        in: query
        name: nome
        type: string
      - description: Filtro por status ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar ingredientes
      tags:
      - Ingredient
  /api/v1/ingredients/{id}:
    get:
      description: Retorna um Ingredient Substance pelo seu ID
      parameters:
      - description: ID do ingrediente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.IngredientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter ingrediente por ID
      tags:
      - Ingredient
...
  /api/v1/vmpp:
    get:
      description: Lista Virtual Medicinal Product Packs com paginacao e filtros
      parameters:
      - default: 20
        description: Limite por pagina
        in: query
        name: limit
        type: integer
      - description: Cursor de paginacao
        in: query
        name: cursor
        type: string
      - description: Filtro por nome (ILIKE)
        in: query
        name: nome
        type: string
      - description: Filtro por status ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar VMPPs
      tags:
      - VMPP
  /api/v1/vmpp/{id}:
    get:
      description: Retorna um Virtual Medicinal Product Pack pelo seu ID
      parameters:
      - description: ID do VMPP
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.VMPPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter VMPP por ID
      tags:
      - VMPP
  /api/v1/vtm:
    get:
      description: Lista Virtual Therapeutic Moieties com paginacao e filtros
      parameters:
      - default: 20
        description: Limite por pagina
        in: query
        name: limit
        type: integer
      - description: Cursor de paginacao
        in: query
        name: cursor
        type: string
      - description: Filtro por nome (ILIKE)
        in: query
        name: nome
        type: string
      - description: Filtro por status ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar VTMs
      tags:
      - VTM
  /api/v1/vtm/{id}:
    get:
      description: Retorna um Virtual Therapeutic Moiety pelo seu ID
      parameters:
      - description: ID do VTM
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.VTMResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_silasrm_api-obm_internal_interface_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter VTM por ID
      tags:
      - VTM
...
securityDefinitions:
  BearerAuth:
    description: 'Tipo: Bearer <token>. Insira o token JWT obtido via /auth/login'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"